Add 3s test option, switch parseInt to parseFloat

+++ src/components/pomodoro.js
@@ -130,11 +130,11 @@ export function createPomodoro(container) {
   }
 
   function getFocusTime() {
-    return parseInt(focusSelect.value) * 60;
+    return parseFloat(focusSelect.value) * 60;
   }
 
   function getBreakTime() {
-    return parseInt(breakSelect.value) * 60;
+    return parseFloat(breakSelect.value) * 60;
   }
 
   function setTimeLeft(secs) {
